Fixed htmx request headers
Necessary to ensure the necessary link:https://htmx.org/reference/#headers[headers] are supplied since we are making an HTTP GET request each time. This extension is unique in that it relies on HTTP GET requests so need to behave like one. The mistage was using `htmx:beforeRequest` instead of `htmx:configRequest` to ensure the proper headers are sent.

I've also made to sure to check that the tag is always a HTML SELECT element so people don't attempt to use other tags which could lead to surprising behavior.

Milestone: patch